+++ web/src/core/components/AppLayout/index.js
