+++ web/src/core/components/LoadingIndicator/index.js
