+++ web/src/core/pages/LoginPage/LoginPage.js
@@ -7,7 +7,7 @@ import { Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 import { LoginForm } from '../../forms';
-import { NavBar } from '../../../components';
+import { NavBar } from '../../components';
 import logo from "../../../assets/images/logo.png";
 
 const LoginPage = () => {
