+++ web/src/entities/communities/pages/CommunitiesPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { faBuilding } from '@fortawesome/free-solid-svg-icons';
 
-import Page from '../../../components/Page/Page';
+import { Page } from '../../../core/components';
 import CommunitiesTable from '../components/CommunitiesTable';
 
 const CommunitiesPage = () => (
