+++ web/src/entities/users/pages/UsersPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { faUsers } from '@fortawesome/free-solid-svg-icons'
 
-import Page from '../../../components/Page/Page';
+import { Page } from '../../../core/components';
 import UsersTable from '../components/UsersTable';
